fix: correcting missing refactor naming
