+++ base/infrastructure/path_management/path_factory/path_creator.py
